fix: picks using config's link to support v2 pools   (#11499)
<!--
Before opening a pull request, please read the [contributing
guidelines](https://github.com/pancakeswap/pancake-frontend/blob/develop/CONTRIBUTING.md)
first
-->


<!-- start pr-codex -->

---

## PR-Codex overview
This PR introduces a new `url` property to the `PicksConfig` type and
updates the `AdPicks` component to utilize this property for linking. It
also removes an unused import and modifies the way the link is
constructed in the rendered output.

### Detailed summary
- Added `url: string` to `PicksConfig` in `types.ts`.
- Updated `AdPicks` component in `AdPicks.tsx` to use `config.url`
instead of `link`.
- Removed the unused import of `useAtomValue` from `jotai`.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->