fix: add harvest button to close position if it have reward (#11599)
<!--
Before opening a pull request, please read the [contributing
guidelines](https://github.com/pancakeswap/pancake-frontend/blob/develop/CONTRIBUTING.md)
first
-->


<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on improving the display of earnings in the
`PositionItem` and enhancing the rendering of `InfinityCLPositionItem`
in the `MyPositions` component.

### Detailed summary
- In `PositionInfo.tsx`, changed the condition from `earningsBusd > 0`
to `earningsAmount > 0` for rendering.
- In `MyPositions.tsx`, modified the return statement to format
`InfinityCLPositionItem` with additional props, including `action`.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->