fix: fix various testnet issue (#11701)
<!--
Before opening a pull request, please read the [contributing
guidelines](https://github.com/pancakeswap/pancake-frontend/blob/develop/CONTRIBUTING.md)
first
-->


<!-- start pr-codex -->

---

## PR-Codex overview
This PR primarily focuses on adding support for testnet scenarios across
multiple files, ensuring that certain functions return appropriate
values when operating in a testnet environment.

### Detailed summary
- In `useCurrencyUsdPrice.ts`, added a check for testnet that returns
`0` if the currency is in a testnet.
- In `getV3CandidatePools.ts`, included a testnet check that returns an
empty array if the currency is in a testnet.
- In `poolQueries.ts`, added a testnet check that returns a fallback
query if in a testnet.
- In `getVerifiedTrade.ts`, modified gas estimation logic to handle
cases when the denominator is `0`, returning `0` for gas estimates
instead.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->