chore: update noAsset condition for walletModal (#11739)
<!--
Before opening a pull request, please read the [contributing
guidelines](https://github.com/pancakeswap/pancake-frontend/blob/develop/CONTRIBUTING.md)
first
-->

<!-- start pr-codex -->

---

## PR-Codex overview
This PR updates the logic for determining if there are no assets in the
`WalletModal` component by including a check for `totalBalanceUsd`. This
enhances the condition to account for cases where the total balance is
zero.

### Detailed summary
- Modified the `noAssets` constant to check if either `topTokens.length`
is zero or `totalBalanceUsd` is zero, in addition to the existing
`!isLoading` condition.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->