fix: xchain refund condition add a partial success case (#11777)
<!--
Before opening a pull request, please read the [contributing
guidelines](https://github.com/pancakeswap/pancake-frontend/blob/develop/CONTRIBUTING.md)
first
-->


<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on enhancing the handling of warning icons in the
`Timeline` component and improving the logic for determining the result
token in the `OrderResultModalContent` component of the swap feature.

### Detailed summary
- Added `mr="2px"` margin to `WarningIcon` for both 'warning' and
'failed' cases in `Timeline/index.tsx`.
- Introduced `lastExecutedCommandIndex` to track the index of the last
executed command in `OrderResultModalContent.tsx`.
- Updated logic to determine `resultTokenAddress` and `resultAmount`
based on the success of the previous step in the bridge status.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->