fix: revert bridge widget version (#11818)
Version 0.9.0 of `canonical-bridge-widget` is currently not showing
routes for any pair.
Revert back to 0.6.1 until it is fixed, and re-enable deBridge referral
code and google analytics afterwards too.

<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on reverting the version of the
`@bnb-chain/canonical-bridge-widget` to `0.6.1`, along with commenting
out several analytics-related code sections in the `CanonicalBridge`
component.

### Detailed summary
- Reverted `@bnb-chain/canonical-bridge-widget` version from `0.9.0` to
`0.6.1` in `package.json`.
- Commented out `createGTMEventListener` initialization in
`CanonicalBridge`.
- Commented out `deBridgeReferralCode` in the configuration.
- Commented out the entire `analytics` section in the `CanonicalBridge`
component.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->