fix: Stablecoin price in useStablecoin hook (#11823)
<!--
Before opening a pull request, please read the [contributing
guidelines](https://github.com/pancakeswap/pancake-frontend/blob/develop/CONTRIBUTING.md)
first
-->

<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on enhancing the `useStablecoinPrice` hook by adding an
`enabled` parameter to control the execution of the price query and
modifying how dependencies are hashed for the atom family.

### Detailed summary
- Removed the import of `CAKE` from `@pancakeswap/tokens`.
- Added an `enabled` property to the `StableCoinPriceParams` interface.
- Updated the condition to return `undefined` in the atom if `currency`
is not provided or `enabled` is `false`.
- Modified the hash generation in the atom family to include the
`enabled` parameter.
- Simplified the `shouldEnabled` condition to exclude checks for
`isCake` and `isStableCoin`.
- Updated the price calculation logic to check for `priceUSD` in
conjunction with `currency`, `stableCoin`, and `shouldEnabled`.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->