fix: replace mint logoURI (#11875)
<!--
Before opening a pull request, please read the [contributing
guidelines](https://github.com/pancakeswap/pancake-frontend/blob/develop/CONTRIBUTING.md)
first
-->


<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on updating the `logoURI` properties of `mintA` and
`mintB` in the `formatPoolData` function if their corresponding entries
are found in the `tokenList`. This ensures that the logo information is
correctly assigned for both tokens.

### Detailed summary
- Added logic to update `mintA.logoURI` with `mintAInTokenList.logoURI`
if both exist.
- Added logic to update `mintB.logoURI` with `mintBInTokenList.logoURI`
if both exist.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->