fix: update link for tips (#11880)
<!-- start pr-codex -->

## PR-Codex overview
This PR focuses on enhancing the user experience in the `CreateClmmPool`
component by updating the note description to include a link to the
documentation for CLMM pools and adding a new entry in the localization
file for "CLMM pools."

### Detailed summary
- Updated the description in the `SubPageNote` component of
`CreateClmmPool` to include a link to the CLMM pools documentation.
- Added the translation entry for "CLMM pools" in `translations.json`.
- Modified the existing translation for the note about advanced users to
separate the guide reference from the pool type.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->