fix: wallet connect issue (#11992)
<!--
Before opening a pull request, please read the [contributing
guidelines](https://github.com/pancakeswap/pancake-frontend/blob/develop/CONTRIBUTING.md)
first
-->

<!-- start pr-codex -->

---

## PR-Codex overview
This PR introduces updates to the `Privy` context and the
`ConnectWalletButton` component in the web application. It adds an
environment file and modifies wallet connection settings, as well as
adjusts the styling for a modal.

### Detailed summary
- Added `.env` file in `apps/web/`
- Updated `externalWallets` in `Privy` context to include
`walletConnect` with `enabled: false`
- Changed modal styling in `ConnectWalletButton` to use `fixed`
positioning and increased `z-index` to `100000000`

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->